backend/python/requirements.txt
# Python dependencies for bulk upload processor

# WebSocket client
python-socketio>=5.10.0

# PostgreSQL database adapter
psycopg2-binary>=2.9.9

# Excel file processing (if not already installed)
openpyxl>=3.1.2
pandas>=2.1.0

# Environment configuration
python-dotenv>=1.0.0

# Logging and utilities
python-dateutil>=2.8.2

